+++ .vscode/settings.json
@@ -1,3 +1,3 @@
 {
-    "python.pythonPath": "/home/maelle/miniconda3/envs/neurodash/bin/python"
+    "python.pythonPath": "/usr/bin/python3"
 }
